Moved sleep timer into its own class.
